Rename handlers to static_handlers
