+++ trustchain-http/src/server.rs
@@ -1,6 +1,6 @@
 // use axum::{routing::get, Router, middleware::{self, Next}, extract::{FromRequest, Request}};
 use crate::middleware::validate_did;
-use crate::{config::HTTPConfig, handlers, issuer, resolver, state::AppState, verifier};
+use crate::{config::HTTPConfig, issuer, resolver, state::AppState, static_handlers, verifier};
 use axum::routing::IntoMakeService;
 use axum::{middleware, routing::get, Router};
 use hyper::server::conn::AddrIncoming;
@@ -30,7 +30,7 @@ impl TrustchainRouter {
     fn new(shared_state: Arc<AppState>) -> Self {
         Self {
             router: Router::new()
-                .route("/", get(handlers::index))
+                .route("/", get(static_handlers::index))
                 .route(
                     "/issuer",
                     get(issuer::TrustchainIssuerHTTPHandler::get_issuer_qrcode),
