+++ trustchain-http/src/static_handlers.rs
